+++ include/trifinger_cameras/pybullet_tricamera_driver.hpp
@@ -10,6 +10,7 @@
 
 #include <robot_interfaces/finger_types.hpp>
 #include <robot_interfaces/sensors/sensor_driver.hpp>
+#include <trifinger_cameras/settings.hpp>
 #include <trifinger_cameras/tricamera_observation.hpp>
 
 namespace trifinger_cameras
@@ -23,7 +24,8 @@ class PyBulletTriCameraDriver : public robot_interfaces::SensorDriver<
 public:
     PyBulletTriCameraDriver(
         robot_interfaces::TriFingerTypes::BaseDataPtr robot_data,
-        bool render_images = true);
+        bool render_images = true,
+        Settings settings = Settings());
 
     /**
      * @brief Get the latest observation from the three cameras
@@ -47,6 +49,9 @@ private:
     //! @brief Last robot time index at which a camera observation was returned.
     //         Needed for time synchronisation.
     time_series::Index last_update_robot_time_index_;
+
+    //! Number of robot time steps after which the next frame should be fetched.
+    int frame_rate_in_robot_steps_;
 };
 
 }  // namespace trifinger_cameras