+++ src/pylon_driver.cpp
@@ -9,7 +9,7 @@
 #include <chrono>
 #include <iostream>
 
-#include <ament_index_cpp/get_package_share_directory.hpp>
+#include <fmt/format.h>
 #include <opencv2/opencv.hpp>
 
 namespace trifinger_cameras
@@ -53,77 +53,87 @@ cv::Mat BGR2BayerBG(const cv::Mat& bgr_image)
     return bayer_img;
 }
 
-PylonDriver::PylonDriver(const std::string& device_user_id,
-                         bool downsample_images)
-    : device_user_id_(device_user_id), downsample_images_(downsample_images)
+void pylon_connect(std::string_view device_user_id,
+                   Pylon::CInstantCamera* camera)
 {
-    try
-    {
-        Pylon::CTlFactory& tl_factory = Pylon::CTlFactory::GetInstance();
-        Pylon::PylonInitialize();
-        Pylon::DeviceInfoList_t device_list;
+    Pylon::PylonInitialize();
+    Pylon::CTlFactory& tl_factory = Pylon::CTlFactory::GetInstance();
+    Pylon::DeviceInfoList_t device_list;
 
-        if (tl_factory.EnumerateDevices(device_list) == 0)
+    if (tl_factory.EnumerateDevices(device_list) == 0)
+    {
+        Pylon::PylonTerminate();
+        throw std::runtime_error("No devices present, please connect one.");
+    }
+
+    Pylon::DeviceInfoList_t::const_iterator device_iterator;
+    if (device_user_id.empty())
+    {
+        device_iterator = device_list.begin();
+
+        fmt::print(
+            "No device ID specified.  Connecting to first camera in the list "
+            "({})\n",
+            device_iterator->GetUserDefinedName());
+        camera->Attach(tl_factory.CreateDevice(*device_iterator));
+    }
+    else
+    {
+        bool found_desired_device = false;
+
+        for (device_iterator = device_list.begin();
+             device_iterator != device_list.end();
+             ++device_iterator)
         {
-            Pylon::PylonTerminate();
-            throw std::runtime_error("No devices present, please connect one.");
+            std::string device_user_id_found(
+                device_iterator->GetUserDefinedName());
+            if (device_user_id == device_user_id_found)
+            {
+                found_desired_device = true;
+                break;
+            }
         }
 
-        Pylon::DeviceInfoList_t::const_iterator device_iterator;
-        if (device_user_id.empty())
+        if (found_desired_device)
         {
-            device_iterator = device_list.begin();
-            camera_.Attach(tl_factory.CreateDevice(*device_iterator));
-            std::cout << "No device ID specified. Creating a camera object "
-                         "with the first device id in the device list."
-                      << std::endl;
+            camera->Attach(tl_factory.CreateDevice(*device_iterator));
         }
         else
         {
-            bool found_desired_device = false;
-
-            for (device_iterator = device_list.begin();
-                 device_iterator != device_list.end();
-                 ++device_iterator)
-            {
-                std::string device_user_id_found(
-                    device_iterator->GetUserDefinedName());
-                if (device_user_id == device_user_id_found)
-                {
-                    found_desired_device = true;
-                    break;
-                }
-            }
-
-            if (found_desired_device)
-            {
-                camera_.Attach(tl_factory.CreateDevice(*device_iterator));
-            }
-            else
-            {
-                Pylon::PylonTerminate();
-                throw std::runtime_error(
-                    "Device id " + device_user_id_ +
-                    " doesn't correspond to any "
-                    "connected devices, please retry with a valid id.");
-            }
-
-            camera_.Open();
-            camera_.MaxNumBuffer = 5;
-
-            set_camera_configuration();
-
-            camera_.StartGrabbing(Pylon::GrabStrategy_LatestImageOnly);
+            Pylon::PylonTerminate();
+            throw std::runtime_error(
+                fmt::format("Device id {} doesn't correspond to any connected "
+                            "devices, please retry with a valid id.",
+                            device_user_id));
         }
+    }
 
+    camera->Open();
+    camera->MaxNumBuffer = 5;
+}
+
+PylonDriver::PylonDriver(const std::string& device_user_id,
+                         bool downsample_images,
+                         Settings settings)
+    : settings_(settings.get_pylon_driver_settings()),
+      downsample_images_(downsample_images)
+{
+    try
+    {
+        pylon_connect(device_user_id, &camera_);
+        // get device user id from camera (useful in case an empty id was
+        // passed, in which case a random camera is connected)
+        device_user_id_ = camera_.GetDeviceInfo().GetUserDefinedName();
+        set_camera_configuration();
+        camera_.StartGrabbing(Pylon::GrabStrategy_LatestImageOnly);
         format_converter_.OutputPixelFormat = Pylon::PixelType_BGR8packed;
     }
     catch (const Pylon::GenericException& e)
     {
-        // convert Pylon exceptions to an std exception, so it is understood by
-        // pybind11
-        throw std::runtime_error("Camera Error (" + device_user_id_ +
-                                 "): " + e.what());
+        // convert Pylon exceptions to an std exception, so it is understood
+        // by pybind11
+        throw std::runtime_error(
+            fmt::format("Camera Error ({}): {}", device_user_id_, e.what()));
     }
 }
 
@@ -150,19 +160,19 @@ CameraObservation PylonDriver::get_observation()
 
         if (ptr_grab_result->GrabSucceeded())
         {
-            // ensure that the actual image size matches with the expected one
+            // ensure that the actual image size matches with the expected
+            // one
             if (ptr_grab_result->GetHeight() / 2 != image_frame.height ||
                 ptr_grab_result->GetWidth() / 2 != image_frame.width)
             {
-                std::stringstream msg;
-                msg << device_user_id_ << ": "
-                    << "Size of grabbed frame (" << ptr_grab_result->GetWidth()
-                    << "x" << ptr_grab_result->GetHeight()
-                    << ") does not match expected size ("
-                    << image_frame.width * 2 << "x" << image_frame.height * 2
-                    << ").";
-
-                throw std::length_error(msg.str());
+                throw std::length_error(
+                    fmt::format("{}: Size of grabbed frame ({}x{}) does not "
+                                "match expected size ({}x{}).",
+                                device_user_id_,
+                                ptr_grab_result->GetWidth(),
+                                ptr_grab_result->GetHeight(),
+                                image_frame.width * 2,
+                                image_frame.height * 2));
             }
 
             if (downsample_images_)
@@ -170,7 +180,8 @@ CameraObservation PylonDriver::get_observation()
                 Pylon::CPylonImage pylon_image_bgr;
                 format_converter_.Convert(pylon_image_bgr, ptr_grab_result);
 
-                // NOTE: the cv::Mat points to the memory of pylon_image_bgr!
+                // NOTE: the cv::Mat points to the memory of
+                // pylon_image_bgr!
                 cv::Mat image_bgr =
                     cv::Mat(ptr_grab_result->GetHeight(),
                             ptr_grab_result->GetWidth(),
@@ -205,16 +216,16 @@ CameraObservation PylonDriver::get_observation()
         }
         else
         {
-            throw std::runtime_error("Failed to grab image from camera " +
-                                     device_user_id_ + ".");
+            throw std::runtime_error(fmt::format(
+                "Failed to grab image from camera '{}'.", device_user_id_));
         }
     }
     catch (const Pylon::GenericException& e)
     {
-        // convert Pylon exceptions to an std exception, so it is understood by
-        // pybind11
-        throw std::runtime_error("Camera Error (" + device_user_id_ +
-                                 "): " + e.what());
+        // convert Pylon exceptions to an std exception, so it is understood
+        // by pybind11
+        throw std::runtime_error(
+            fmt::format("Camera Error ({}): {}", device_user_id_, e.what()));
     }
 
     return image_frame;
@@ -248,16 +259,7 @@ cv::Mat PylonDriver::downsample_raw_image(const cv::Mat& image)
 
 void PylonDriver::set_camera_configuration()
 {
-    const std::string filename =
-        ament_index_cpp::get_package_share_directory("trifinger_cameras") +
-        "/config/pylon_camera_settings.txt";
-
     Pylon::CFeaturePersistence::Load(
-        filename.c_str(), &camera_.GetNodeMap(), true);
-
-    // Use the following command to generate a config file with the current
-    // settings:
-    // Pylon::CFeaturePersistence::Save("/tmp/camera_settings.txt",
-    //                                 &camera_.GetNodeMap());
+        settings_->pylon_settings_file.c_str(), &camera_.GetNodeMap(), true);
 }
 }  // namespace trifinger_cameras